Henchmen hire and quit changes the neutral population; convert henchmen quantities type from int to long

+++ src/wmd-core/Commands/HireHenchmenResult.cs
@@ -14,7 +14,7 @@ namespace WMD.Game.Commands
         /// <param name="updatedGameState">The updated <see cref="GameState"/> resulting from this action.</param>
         /// <param name="playerIndex">The index of the <see cref="Player"/> whose action this is the result of.</param>
         /// <param name="henchmenHired">The number of henchmen successfully hired.</param>
-        public HireHenchmenResult(GameState updatedGameState, int playerIndex, int henchmenHired) : base(updatedGameState, playerIndex)
+        public HireHenchmenResult(GameState updatedGameState, int playerIndex, long henchmenHired) : base(updatedGameState, playerIndex)
         {
             HenchmenHired = henchmenHired;
         }
@@ -22,6 +22,6 @@ namespace WMD.Game.Commands
         /// <summary>
         /// Gets the number of henchmen successfully hired.
         /// </summary>
-        public int HenchmenHired { get; init; }
+        public long HenchmenHired { get; init; }
     }
 }
